Changed "projectsbrowser" top padding
By adding a body-wide padding to account for the navbar, we can remove
the projectsbrowser one, but it makes the images shove right up under
the navbar. For now, I've put a projectsbrowser padding.

It may be better to put the padding on the images and remove the
section's padding? Or maybe we can do both, but that's for later.